Use tuareg-match-clause-indent when indenting arguments.
In the case of a match clause that is a function call, when the
function appears on the same line as the clause, and when the arguments
appear on separate lines, the indentation looks like

    match () with
    | _ -> some_function
       argument1
      argument2

with tuareg-match-clause-indent equal to 1. This commit fixes the
problem by taking tuareg-match-clause-indent into account when indenting
arguments.