Various indentation improvements.  Add regression test
* tuareg.el: Various indentation improvements.
(tuareg-smie--|-or-p): New function.
(tuareg-smie-backward-token): Refine detection of |-or (Issue#76).
(tuareg-smie-rules): Refine rules for datatype | indentation.
Add support for the new :close-all kind.  Fix alignment of "with" used
in type definitions.  Improve virtual indentation rule for open parens.
(tuareg-smie--if-then-hack): Align | with the end of the
corresponding "with".

* Makefile (%.test): Fix rule so tuareg_indent is found.
(indent-test): New rule.

* indent-test.ml: New file, taken from sample.ml.

* .gitignore: Add *.test.