Fix insertion of backslashes in non-"" strings
This fixes the bug where backslashes are inserted on `RET` in `{| string |}` literals and embedded minor modes like [smerge-mode](https://github.com/jwiegley/emacs-release/blob/master/lisp/vc/smerge-mode.el), where they are unwanted.

I wrote this as a patch in my init.el and when I went to submit it here I noticed #54 but that PR is enormous and over a year old and seems to be held up indefinitely. I propose this tiny patch as a way to fix the bug for now. Maybe eventually some better solution can be put in place, but this annoying little bug should be patched as soon as possible in the mean time.