Look for the type immediately after the let-binding
Example: The "ret" type  should be highlighted in

let f x : ret = body
let f (x) : ret = body

but the search should not extend to the rest of the line.  In

let f x = something ~l:(fun x -> x = y)

the previous regular expression was matching "(fun x -> x " as the
type because it was searched with the end of the line as a
bound (instead of the first "=").