Make beginning-of-defun (C-M-a) repeatable
Ensure that `tuareg-beginning-of-defun` right at the beginning of a
defun moves point to the beginning of the previous defun rather than
being a no-op (issue #236). Also make the return value is correct for
a `beginning-of-defun-function`, as well as the semantics for negative
arguments.

Add tests, which are run automatically as part of Travis CI, or
manually by `make check-ert` or M-x ert.