Let declarative `and` begin a defun
Since each `and` hitched to a `type` or declarative `let` produces a
definition, consider them defuns in their own right. Previously, an
entire `let`...`and`... or `type`...`and`... chain was a single defun,
which made movement-by-defun operations less useful.

To get an acceptable interactive latency, we memoise the
calls (smie-forward-sexp "and") and (smie-backward-sexp "and"),
since they are called repeatedly during movement-by-defun.