Fix indentation and paragraph-fill inside comments
Inside comments, explicit indentation as well as the automatic
indentation with electric-indent-mode active (just pressing RET, in
Emacs 28) now use the indentation of the previous nonempty line, even
when that line is on the same line as the initial `(*`.

Filling paragraphs with M-q inside a comment now only fills a single
paragraph instead of the entire comment. On an empty line, the
preceding paragraph is used.

Inside doc comments, lines starting with @-tags (like `@param`) start
a paragraph, and subsequent lines are indented to the text after the
tag.

Fixes #213.