Prevent Invalid search bound error on long patterns
When `tuareg--pattern-pre-form-let' is invoked at the beginning of a
let binding whose LHS is very long (> 800 characters), it can signal
the following error:

    Invalid search bound (wrong side of point)

This happens when a balanced set of parentheses includes the limit
position, so `forward-list' skips past it and thereby causes point to
exceed limit.

Apply the same fix to `tuareg--pattern-pre-form-fun' by analogy.
`tuareg--pattern-equal-matcher' is already correct.